+++ src/components/CurriculumComponents/UnitsTab/UnitsTab.test.tsx
@@ -86,6 +86,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
           subject_parent: null,
           subject_parent_slug: null,
           subject_slug: "english",
+          tags: null,
           threads: [
             {
               title: "Developing grammatical knowledge",
@@ -121,6 +122,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
           subject_slug: "english",
           tier: null,
           tier_slug: null,
+          tags: null,
           threads: [
             {
               title: "Aspiration",
@@ -217,6 +219,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
           subject_parent: null,
           subject_parent_slug: null,
           subject_slug: "english",
+          tags: null,
           threads: [],
           tier: null,
           tier_slug: null,
@@ -263,6 +266,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
           subject_parent: null,
           subject_parent_slug: null,
           subject_slug: "english",
+          tags: null,
           threads: [],
           tier: null,
           tier_slug: null,
@@ -350,6 +354,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
           subject_parent: "Science",
           subject_parent_slug: "science",
           subject_slug: "combined-science",
+          tags: null,
           threads: [],
           tier: null,
           tier_slug: null,
@@ -377,6 +382,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
           subject_parent: "Science",
           subject_parent_slug: "science",
           subject_slug: "physics",
+          tags: null,
           threads: [],
           tier: null,
           tier_slug: null,
@@ -428,6 +434,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
           subject_parent: null,
           subject_parent_slug: null,
           subject_slug: "english",
+          tags: null,
           threads: [],
           tier: null,
           tier_slug: null,
@@ -457,6 +464,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
           subject_slug: "english",
           tier: null,
           tier_slug: null,
+          tags: null,
           threads: [],
           title: "’A Superhero Like You!’: Reading and Writing",
           unit_options: [],
@@ -479,6 +487,62 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
     await waitFor(async () => {
       unitCards = await findAllByTestId("unit-card");
       expect(unitCards).toHaveLength(1);
+    });
+  });
+
+  test("user can filter units by discipline", async () => {
+    const data = curriculumUnitsTabFixture();
+    const { findAllByTestId, queryAllByTestId } = render(
+      <UnitsTab data={data} examboardSlug={null} />,
+    );
+
+    const yearOptions = (await queryAllByTestId(
+      "year-radio",
+    )) as HTMLInputElement[];
+
+    const year10Option = yearOptions.find((option) => option.value === "10");
+    if (!year10Option) {
+      throw new Error("No year 10 option found");
+    }
+
+    const disciplineButtons = await findAllByTestId("discipline-button");
+
+    if (
+      !disciplineButtons[0] ||
+      !disciplineButtons[1] ||
+      !disciplineButtons[2]
+    ) {
+      throw new Error("Missing second subject button");
+    }
+
+    expect(disciplineButtons[0]).toHaveTextContent("All");
+    expect(disciplineButtons[1]).toHaveTextContent("Biology");
+    expect(disciplineButtons[2]).toHaveTextContent("Physics");
+
+    // Check we only have 2 units for year 10 to start with.
+    await userEvent.click(year10Option);
+    let unitCards;
+    await waitFor(async () => {
+      unitCards = await findAllByTestId("unit-card");
+      expect(unitCards).toHaveLength(2);
+    });
+
+    // Check we have 1 unit after clicking the Biology button.
+    await userEvent.click(disciplineButtons[1]);
+    await waitFor(async () => {
+      unitCards = await findAllByTestId("unit-card");
+      expect(unitCards).toHaveLength(1);
+      if (!data.units[1]?.title) {
+        throw new Error("No unit title in fixture");
+      }
+      expect(unitCards[0]).toHaveTextContent(data.units[1]?.title);
+    });
+
+    // Check we have 2 units after clicking the All button.
+    await userEvent.click(disciplineButtons[0]);
+    await waitFor(async () => {
+      unitCards = await findAllByTestId("unit-card");
+      expect(unitCards).toHaveLength(2);
     });
   });
 
@@ -505,6 +569,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
           subject_parent: "Science",
           subject_parent_slug: "science",
           subject_slug: "combined-science",
+          tags: null,
           threads: [],
           tier: "Foundation",
           tier_slug: "foundation",
@@ -519,6 +584,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
           connection_prior_unit_title: null,
           domain: null,
           domain_id: null,
+          tags: null,
           threads: [],
           examboard: null,
           examboard_slug: null,
@@ -577,6 +643,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
           connection_prior_unit_title: null,
           domain: null,
           domain_id: null,
+          tags: null,
           threads: [],
           examboard: null,
           examboard_slug: null,
@@ -641,6 +708,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
           subject_parent: null,
           subject_parent_slug: null,
           subject_slug: "english",
+          tags: null,
           threads: [],
           tier: null,
           tier_slug: null,
@@ -698,6 +766,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
         subject_parent: "Science",
         subject_parent_slug: "science",
         subject_slug: "combined-science",
+        tags: null,
         threads: [],
         tier: "Foundation",
         tier_slug: "foundation",
@@ -730,6 +799,7 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
         subject_parent: "Science",
         subject_parent_slug: "science",
         subject_slug: "combined-science",
+        tags: null,
         threads: [],
         tier: "Foundation",
         tier_slug: "foundation",